// frontend/src/api/models/personal-info-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Resume Car Builder API
 * An API for Resume Car Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PersonalInfoVM
 */
export interface PersonalInfoVM {
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    fullName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PersonalInfoVM
     */
    personImageURL?: string | null;
}