// frontend/src/api/models/resume-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Resume Car Builder API
 * An API for Resume Car Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ResumeVM
 */
export interface ResumeVM {
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    title?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ResumeVM
     */
    creationTime?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ResumeVM
     */
    lastModified?: Date;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    userId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    userFullNames?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    personalInfoId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    templateId?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ResumeVM
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ResumeVM
     */
    carInfoId?: string | null;
}