// frontend/src/api/models/template-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Resume Car Builder API
 * An API for Resume Car Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface TemplateVM
 */
export interface TemplateVM {
    /**
     * 
     * @type {string}
     * @memberof TemplateVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TemplateVM
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TemplateVM
     */
    content?: string | null;
}